rename py files and fix references
