+++ plotmanager/plottype/anim_plot.py
