+++ plotmanager/plottype/cluster_map.py
