+++ plotmanager/plottype/image_stack.py
