+++ plotmanager/plottype/protein_track.py
